better error messages in docs

+++ backend/handlers/pair_handler.py
@@ -1,4 +1,4 @@
-from fastapi import APIRouter, Request, UploadFile, File, Form
+from fastapi import APIRouter, Request, UploadFile, File, Form, status
 from fastapi.responses import JSONResponse
 from pydantic import BaseModel
 from sqlalchemy import select
@@ -18,7 +18,7 @@ async def get_pairs(request: Request) -> List[PairResponse]:
     with Session(request.app.state.db) as session:
         return session.execute(select(Pair)).scalars().all()
 
-@router.get("/{patient_id}")
+@router.get("/{patient_id}", responses={404: {"description": "Patient not found"}})
 async def get_pairs_by_patient_id(req: Request, patient_id: int) -> List[PairResponse]:
     """
     Get all pairs for a patient
@@ -26,11 +26,11 @@ async def get_pairs_by_patient_id(req: Request, patient_id: int) -> List[PairRes
     with Session(req.app.state.db) as session:
         patient = session.execute(select(Patient).filter(Patient.id == patient_id)).scalars().first()
         if not patient:
-            return JSONResponse(status_code=404, content={"message": "Patient not found"})
+            return JSONResponse(status_code=status.HTTP_404_NOT_FOUND, content={"message": "Patient not found"})
 
         return session.execute(select(Pair).filter(Pair.patient_id == patient_id)).scalars().all()
 
-@router.get("/{patient_id}/{filename}")
+@router.get("/{patient_id}/{filename}", responses={404: {"description": "Patient or File not found"}})
 async def get_file(req: Request, patient_id: int, filename: str) -> bytes:
     """
     Get a file for a patient
@@ -38,12 +38,12 @@ async def get_file(req: Request, patient_id: int, filename: str) -> bytes:
     with Session(req.app.state.db) as session:
         patient = session.execute(select(Patient).filter(Patient.id == patient_id)).scalars().first()
         if not patient:
-            return JSONResponse(status_code=404, content={"message": "Patient not found"})
+            return JSONResponse(status_code=status.HTTP_404_NOT_FOUND, content={"message": "Patient not found"})
     bucket = req.app.state.gcs_client.get_bucket(patient.id_string)
 
     filenames = [blob.name for blob in bucket.list_blobs()]
     if filename not in filenames:
-        return JSONResponse(status_code=404, content={"message": "File not found"})
+        return JSONResponse(status_code=status.HTTP_404_NOT_FOUND, content={"message": "File not found"})
 
     blob = bucket.blob(filename)
     return blob.download_as_string()
@@ -60,7 +60,7 @@ def upload_file(gcs_client, patient_id, file):
     blob.upload_from_file(file.file)
     return file.filename
 
-@router.post("/{patient_id}")
+@router.post("/{patient_id}", responses={404: {"description": "Patient not found"}})
 async def make_pair(req: Request, patient_id: int, object_one_value: Union[str, UploadFile] = Form(...), object_two_value: Union[str, UploadFile] = Form(...)) -> PairResponse:
     """
     Create a pair for a patient
@@ -68,7 +68,7 @@ async def make_pair(req: Request, patient_id: int, object_one_value: Union[str, 
     with Session(req.app.state.db) as session:
         patient = session.execute(select(Patient).filter(Patient.id == patient_id)).scalars().first()
         if not patient:
-            return JSONResponse(status_code=404, content={"message": "Patient not found"})
+            return JSONResponse(status_code=status.HTTP_404_NOT_FOUND, content={"message": "Patient not found"})
     
     if isinstance(object_one_value, str):
         type_one = "string"
@@ -95,7 +95,7 @@ def delete_file(gcs_client, patient_id, filename):
     blob = bucket.blob(filename)
     blob.delete()
 
-@router.delete("/{pair_id}")
+@router.delete("/{pair_id}", responses={404: {"description": "Patient or Pair not found"}})
 async def delete_pair(req: Request, pair_id: int) -> PairResponse:
     """
     Delete a pair by ID
@@ -103,8 +103,10 @@ async def delete_pair(req: Request, pair_id: int) -> PairResponse:
     with Session(req.app.state.db) as session:
         patient = session.execute(select(Patient).filter(Patient.id == pair.patient_id)).scalars().first()
         if not patient:
-            return JSONResponse(status_code=404, content={"message": "Patient not found"})
+            return JSONResponse(status_code=status.HTTP_404_NOT_FOUND, content={"message": "Patient not found"})
         pair = session.execute(select(Pair).filter(Pair.id == pair_id)).scalars().first()
+        if not pair:
+            return JSONResponse(status_code=status.HTTP_404_NOT_FOUND, content={"message": "Pair not found"})
         if pair.object_one_type == "file":
             delete_file(req.app.state.gcs_client, patient.id_string, pair.object_one_value)
         if pair.object_two_type == "file":
